fix: Tag template  - failed to use
Added a function for sidebar object tags
