+++ src/js/changeBgPage.js
@@ -1,4 +1,4 @@
-export function changeBgPage(body, paginationPages, paginationBtns) {
+export const  changeBgPage = (body, paginationPages, paginationBtns) => {
 
 
   const colorArray = ['#c8dcdb', '#9ab5b5', '#a1acab', '#f7e7d2', '#e2ba65', '#f1d7da'];
