+++ src/js/index.js
@@ -1,13 +1,18 @@
 'use strict;'
-import { sortMenu } from './sortMenu';
-import { createCards } from './createАrrayСards';
+import { modalSortMenu } from './modalSortMenu';
+import { createCard } from './createСard';
 import { changeBgPage } from './changeBgPage';
+
+// import { createCard2 } from './createCard2';
 
 const body = document.body;
 const cardsContainer = document.querySelector('.js-cards-container');
 const paginationPages = document.querySelectorAll('.js-page-btn');
 const paginationBtns = document.querySelectorAll('.js-pagination-arrow');
 
-sortMenu(body);
-createCards(cardsContainer)
+modalSortMenu(body);
+createCard(cardsContainer);
+
+// createCard2(cardsContainer);
+
 changeBgPage(body, paginationPages, paginationBtns);
